style: seperate minimap-full_minimap
