+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "files.associations": {
+        "array": "c",
+        "string": "c",
+        "string_view": "c",
+        "ranges": "c",
+        "span": "c"
+    }
+}
